requirements/python.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
Cerberus==1.3.1
chardet==3.0.4
idna==2.8
multidict==4.5.2
psycopg2-binary
yarl==1.3.0
